Added crit calculation for enemies

+++ Assets/Scripts/Stats/BaseStats.cs
@@ -15,9 +15,9 @@ public class BaseStats : MonoBehaviour
     public int health;
     public int shield;
     public int attack;
-    public int attackSpeed;
+    public float attackSpeed;
     public int critRate;
-    public int critDamage;
+    public float critDamage;
     public bool isImmune = false;
 
     private Coroutine immuneRoutine;
